backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
Werkzeug==2.3.7
eventlet==0.33.3
flask==2.3.3
flask-socketio==5.3.6
pymongo==4.6.1
bcrypt==4.0.1
pyjwt==2.8.0
python-dotenv==1.0.0
eventlet==0.33.3